gateway: split stream contracts and the dialect dispatch seam for the line budget
Mechanical module splits with no behavior change, unblocking the 999-line
gate after main's growth landed in the same files. The provider-neutral
stream outcome contracts (GatewayUsage, GatewayEventKind, GatewayEvent,
GatewayFailureClass, GatewayFailure) move to
exp/runtime/gateway/stream_contracts.py, and the wire dispatch seam
(dialect_stream_payload, the tool-result image degrade, and the
service-tier dialect set) moves to
exp/runtime/models/providers/dialect_dispatch.py. Both origins re-export
every moved name, so import paths are unchanged.
