+++ exp/runtime/gateway/stream_contracts_test.py
@@ -1,0 +1,2 @@
+"""Inline tests for the stream outcome contracts (currently exercised through
+``contracts_test`` and every consumer suite)."""
