+++ exp/runtime/models/providers/streaming_requests.py
@@ -5,17 +5,30 @@ from __future__ import annotations
 import logging
 from collections.abc import Sequence
 from typing import TYPE_CHECKING
-from urllib.parse import urlsplit
 
-from exp.common.core.artifacts import JsonObject
 from exp.runtime.gateway.contracts import (
-    GatewayApiSurface,
-    GatewayMessage,
     GatewayNamedToolChoice,
     GatewayRequest,
 )
+from exp.runtime.models.providers.dialect_dispatch import (
+    SERVICE_TIER_DIALECTS as SERVICE_TIER_DIALECTS,
+)
+from exp.runtime.models.providers.dialect_dispatch import (
+    TOOL_RESULT_IMAGE_DROP_DISCLOSURE as TOOL_RESULT_IMAGE_DROP_DISCLOSURE,
+)
+from exp.runtime.models.providers.dialect_dispatch import (
+    TOOL_RESULT_IMAGE_PLACEHOLDER as TOOL_RESULT_IMAGE_PLACEHOLDER,
+)
+from exp.runtime.models.providers.dialect_dispatch import (
+    dialect_stream_payload as dialect_stream_payload,
+)
+from exp.runtime.models.providers.dialect_dispatch import (
+    fireworks_continuation_required as fireworks_continuation_required,
+)
+from exp.runtime.models.providers.dialect_dispatch import (
+    strip_tool_result_images as strip_tool_result_images,
+)
 from exp.runtime.models.providers.errors import (
-    ProviderCapabilityError,
     ProviderParameterError,
     UnsupportedReasoningEffortError,
 )
@@ -32,13 +45,19 @@ from exp.runtime.models.providers.generation_parameter_validation import (
     require_route_numeric_parameter as _require_route_numeric_parameter,
 )
 from exp.runtime.models.providers.messages_payloads import (
-    anthropic_messages_stream_payload,
-    bedrock_converse_stream_payload,
-    gemini_generate_content_stream_payload,
+    anthropic_messages_stream_payload as anthropic_messages_stream_payload,
+)
+from exp.runtime.models.providers.messages_payloads import (
+    bedrock_converse_stream_payload as bedrock_converse_stream_payload,
+)
+from exp.runtime.models.providers.messages_payloads import (
+    gemini_generate_content_stream_payload as gemini_generate_content_stream_payload,
 )
 from exp.runtime.models.providers.openai_payloads import (
-    openai_compatible_stream_payload,
-    openai_responses_stream_payload,
+    openai_compatible_stream_payload as openai_compatible_stream_payload,
+)
+from exp.runtime.models.providers.openai_payloads import (
+    openai_responses_stream_payload as openai_responses_stream_payload,
 )
 from exp.runtime.models.providers.reasoning_compat import (
     REASONING_EFFORTS,
@@ -52,46 +71,6 @@ if TYPE_CHECKING:
 _logger = logging.getLogger(__name__)
 
 _ANTHROPIC_REQUIRED_MAX_TOKENS_DEFAULT = 4096
-
-TOOL_RESULT_IMAGE_DROP_DISCLOSURE = "messages.content.tool_result.image->placeholder"
-"""Disclosure recorded when tool-result images degrade to placeholder text.
-
-A tool screenshot is baked into the caller's conversation history: rejecting
-it wedges every later turn of a multi-turn session, which is strictly worse
-than a disclosed degrade. Top-level user images keep the fail-closed contract
-because the caller can re-send those differently.
-"""
-
-TOOL_RESULT_IMAGE_PLACEHOLDER = "[image omitted: this model route cannot carry tool-result images]"
-"""Text substituted for each dropped tool-result image, in block position."""
-
-
-def strip_tool_result_images(
-    messages: tuple[GatewayMessage, ...],
-) -> tuple[GatewayMessage, ...] | None:
-    """Replace tool-message image parts with positional placeholder text.
-
-    Args:
-        messages: The request's canonical messages.
-
-    Returns:
-        The degraded messages, or ``None`` when no tool message carries an
-        image (nothing to strip).
-    """
-    if not any(message.role == "tool" and message.images for message in messages):
-        return None
-    out: list[GatewayMessage] = []
-    for message in messages:
-        if message.role != "tool" or not message.images:
-            out.append(message)
-            continue
-        content = "".join(
-            part.text if part.kind == "text" else TOOL_RESULT_IMAGE_PLACEHOLDER
-            for part in message.content_parts
-        )
-        out.append(message.model_copy(update={"content": content, "content_parts": ()}))
-    return tuple(out)
-
 
 GATEWAY_GENERATION_PARAMETER_CONTRACT_VERSION = 2
 """Version of the route admission and provider wire-translation contract."""
@@ -122,20 +101,6 @@ def _serves_reasoning_summary(profile: GatewayWireProfile) -> bool:
     return profile.supports_reasoning and profile.dialect in _REASONING_SUMMARY_DIALECTS
 
 
-def _fireworks_continuation_required(profile: GatewayWireProfile, request: GatewayRequest) -> bool:
-    """Return whether a Fireworks Responses turn can emit an unretained tool call."""
-    return (
-        request.surface == GatewayApiSurface.RESPONSES
-        and (urlsplit(profile.url).hostname or "").lower() == "api.fireworks.ai"
-        and bool(request.tools)
-        and request.tool_choice != "none"
-    )
-
-
-SERVICE_TIER_DIALECTS = frozenset({"openai_responses", "openai_compatible"})
-"""Wire dialects with a request field that preserves the caller's service tier."""
-
-
 def _anthropic_reasoning_disengaged(request: GatewayRequest) -> bool:
     """Whether an Anthropic dispatch will send no extended-thinking budget.
 
@@ -151,123 +116,6 @@ def _anthropic_reasoning_disengaged(request: GatewayRequest) -> bool:
     thinking_on = config is not None and config.get("type") in {"enabled", "adaptive"}
     effort_on = request.reasoning_effort is not None and request.reasoning_effort != "none"
     return not thinking_on and not effort_on
-
-
-def dialect_stream_payload(
-    profile: GatewayWireProfile,
-    provider_request: GatewayRequest,
-) -> JsonObject:
-    """Build the provider wire payload for one resolved wire profile.
-
-    Args:
-        profile: The resolved connection's wire profile.
-        provider_request: Canonical request forced into streaming mode.
-
-    Returns:
-        The exact JSON payload the gateway sends upstream for this dialect.
-
-    Raises:
-        ProviderCapabilityError: The request uses a capability this dialect
-            cannot preserve.
-    """
-    if _fireworks_continuation_required(profile, provider_request):
-        require_responses_continuation_channel(provider_request)
-    if provider_request.service_tier is not None and profile.dialect not in SERVICE_TIER_DIALECTS:
-        # A processing-tier hint changes pricing and latency semantics, so a
-        # dialect with no wire field for it declines instead of dropping the
-        # field silently: admission then prefers a tier-preserving rung and
-        # otherwise retries with the disclosed drop in capability_policy.
-        raise ProviderCapabilityError(capability="service_tier")
-    required_reasoning_effort = (
-        profile.reasoning_effort if profile.reasoning_effort_required else None
-    )
-    if profile.dialect == "openai_responses":
-        return openai_responses_stream_payload(
-            profile.model_id,
-            provider_request,
-            supports_temperature=profile.supports_temperature,
-            supports_top_p=(
-                profile.supports_temperature
-                if profile.supports_top_p is None
-                else profile.supports_top_p
-            ),
-            supports_top_k=profile.supports_top_k,
-            supports_logprobs=profile.supports_logprobs,
-            supports_reasoning=profile.supports_reasoning,
-            reasoning_effort=required_reasoning_effort,
-            sampling_requires_reasoning_none=profile.sampling_requires_reasoning_none,
-            forwards_service_tier=profile.billing_customer_managed,
-        )
-    if profile.dialect == "anthropic_messages":
-        return anthropic_messages_stream_payload(
-            profile.model_id,
-            provider_request,
-            supports_temperature=profile.supports_temperature,
-            supports_top_p=(
-                profile.supports_temperature
-                if profile.supports_top_p is None
-                else profile.supports_top_p
-            ),
-            supports_top_k=profile.supports_top_k,
-            supports_logprobs=profile.supports_logprobs,
-            supports_reasoning=profile.supports_reasoning,
-            reasoning_effort=required_reasoning_effort,
-        )
-    if profile.dialect == "gemini_generate_content":
-        return gemini_generate_content_stream_payload(
-            profile.model_id,
-            provider_request,
-            supports_temperature=profile.supports_temperature,
-            supports_top_p=(
-                profile.supports_temperature
-                if profile.supports_top_p is None
-                else profile.supports_top_p
-            ),
-            supports_top_k=profile.supports_top_k,
-            supports_logprobs=profile.supports_logprobs,
-            supports_reasoning=profile.supports_reasoning,
-            reasoning_effort=required_reasoning_effort,
-        )
-    if profile.dialect == "bedrock_converse_stream":
-        return bedrock_converse_stream_payload(
-            profile.model_id,
-            provider_request,
-            supports_temperature=profile.supports_temperature,
-            supports_top_p=(
-                profile.supports_temperature
-                if profile.supports_top_p is None
-                else profile.supports_top_p
-            ),
-            supports_top_k=profile.supports_top_k,
-            supports_logprobs=profile.supports_logprobs,
-        )
-    if profile.dialect == "openai_compatible":
-        if profile.fireworks_reasoning_route_sha256 is not None:
-            require_responses_continuation_channel(provider_request)
-        return openai_compatible_stream_payload(
-            profile.model_id,
-            provider_request,
-            token_limit_key=profile.token_limit_key,
-            supports_temperature=profile.supports_temperature,
-            supports_top_p=(
-                profile.supports_temperature
-                if profile.supports_top_p is None
-                else profile.supports_top_p
-            ),
-            supports_top_k=profile.supports_top_k,
-            supports_frequency_penalty=profile.supports_frequency_penalty,
-            supports_presence_penalty=profile.supports_presence_penalty,
-            supports_logprobs=profile.supports_logprobs,
-            supports_reasoning=profile.supports_reasoning,
-            reasoning_wire_format=profile.reasoning_wire_format,
-            reasoning_effort=required_reasoning_effort,
-            sampling_requires_reasoning_none=profile.sampling_requires_reasoning_none,
-            fireworks_reasoning_route_sha256=profile.fireworks_reasoning_route_sha256,
-            hunyuan_reasoning_route_sha256=profile.hunyuan_reasoning_route_sha256,
-            reasoning_output_exposed=profile.reasoning_output_exposed,
-            forwards_service_tier=profile.billing_customer_managed,
-        )
-    raise ProviderCapabilityError(capability=f"wire_dialect:{profile.dialect}")
 
 
 def _mid_conversation_system_present(request: GatewayRequest) -> bool:
@@ -311,7 +159,7 @@ def route_generation_parameter_requests(
     if not profiles:
         raise ValueError("generation parameter shaping requires at least one wire profile")
     for profile in profiles:
-        if _fireworks_continuation_required(profile, request):
+        if fireworks_continuation_required(profile, request):
             require_responses_continuation_channel(request)
 
     ignored = list(request.ignored_parameters)
